# Optimize like function for exact, prefix and suffix patterns