--- a/velox/benchmarks/basic/CMakeLists.txt
+++ b/velox/benchmarks/basic/CMakeLists.txt
@@ -57,3 +57,7 @@ target_link_libraries(velox_benchmark_feature_normalization
 add_executable(velox_benchmark_basic_preproc Preproc.cpp)
 target_link_libraries(velox_benchmark_basic_preproc ${velox_benchmark_deps}
                       velox_functions_prestosql velox_vector_test_lib)
+
+add_executable(velox_like_functions_benchmark LikeFunctionsBenchmark.cpp)
+target_link_libraries(velox_like_functions_benchmark ${velox_benchmark_deps}
+                      velox_functions_lib velox_tpch_gen velox_vector_test_lib)