Optimize like function for the following common cases:
1) Patterns with wildcard characters only:
  a) Patterns with only '\_': Match length of input with number of '\_'
      characters.
  b) Patterns with atleast one '%': Check if input length is atleast equal to
       the number of '_' characters.
2) Constant pattern search: Simple string comparison character by
    character.
3) Pattern has constant prefix followed by one or more %: Match the prefix
     only.
4) Pattern has constant suffix preceded by one or more %: Match the suffix
     only.

These cases are most common in TPC-H queries. When an escape character is
involved in the like function, the default Re2 matcher is used. This is
currently a limitation and needs further examination.

The benchmarks obtained with the changes, for inputs generated by vector fuzzer,
and Tpch queries with patterns matching the above, are as follows:
```
| No. of rows | PatternKind / Query Number | With PR (time/row in ns) | Without PR (time/row in ns) | % Change |
|-------------|----------------------------|--------------------------|-----------------------------|----------|
|    10000    |          kExactlyN         |           0.73           |            700.93           |   99.9   |
|             |          kAtLeastN         |           0.701          |            689.71           |   99.9   |
|             |           kFixed           |           3.66           |            144.7            |   97.47  |  
|             |           kPrefix          |           3.26           |            146.39           |   97.77  | 
|             |           kSuffix          |            4.4           |            670.13           |   99.34  | 
|             |              2             |           6.43           |            436.12           |   98.53  |  
|             |             14             |           5.16           |            198.94           |   97.41  |   
|             |           16 Part          |           6.36           |            157.5            |   95.96  |
|             |             20             |            3.9           |            150.26           |   97.4   |    
|    50000    |          kExactlyN         |           0.623          |            695.33           |   99.91  |      
|             |          kAtLeastN         |           0.592          |            690.4            |   99.91  |
|             |           kFixed           |           3.62           |            144.29           |   97.49  |
|             |           kPrefix          |           3.59           |            145.88           |   97.54  |
|             |           kSuffix          |           5.06           |            671.87           |   99.25  |
|             |              2             |           6.19           |            438.01           |   98.59  |   
|             |             14             |           5.41           |            200.07           |   97.3   |    
|             |           16 Part          |           7.39           |            157.33           |   95.3   | 
|             |             20             |           4.46           |            150.76           |   97.04  |
```