@karteekmurthys 2022-06-16
> ideas from DuckDb implementation

Detail it out, what is the idea and how is it optimized with some example.