@pramodsatya 2022-08-30
Thanks for the inputs @mbasmanova. The benchmark changes are in a separate PR (#1838), shall I integrate it with these changes? The following benchmarks were obtained:
```
| 4000 runs (time in minutes) |                    |                       |                 |
|:---------------------------:|:------------------:|:---------------------:|:---------------:|
|            Query            | With optimizations | Without optimizations | Improvement (%) |   
|              q2             |        0.93        |          1.15         |   19.13   |   
|              q9             |        1.17        |          1.39         |   15.82   |   
|             q13             |        2.14        |          2.13         |  -0.46  |   
|             q14             |        0.852       |          0.76         |   -12.10  |   
|            q16_1            |        0.846       |         0.843         |  -0.35  |
|            q16_2            |         2.1        |          2.12         |   0.94  | 
|             q20             |        0.775       |         0.795         |    2.51   |
```