@mbasmanova 2022-08-30
@pramodsatya Re: benchmarks - I was hoping we can create a standalone benchmark that runs LIKE operator for a few vectors and compares the performance of RE2-based implementation with optimized implementation. 

Query-wide benchmarks are also useful, but would be outside of the scope of this PR. BTW, I see that some queries regressed with the optimization. That's surprising. Do you know why is that?