@pramodsatya 2022-09-01
The benchmark is added and the following numbers were obtained:
```
| No. of runs | Pattern Kind  | With optimization PR (s) | Without optimization PR (s) | % Change |
|-------------|---------------|--------------------------|-----------------------------|----------|
|         500 | Wildcard only |                    14.76 |                       15.13 |     2.45 |
|             | Fixed         |                      4.2 |                        4.17 |    -0.72 |
|             | Prefix        |                     9.92 |                        12.5 |    20.64 |
|             | Suffix        |                     8.44 |                        8.88 |     4.95 |
|        1000 | Wildcard only |                    24.44 |                       23.73 |    -2.99 |
|             | Fixed         |                     6.16 |                        6.49 |     5.08 |
|             | Prefix        |                    14.24 |                       18.08 |    21.24 |
|             | Suffix        |                    13.11 |                       14.09 |     6.96 |
|        5000 | Wildcard only |                    109.2 |                       114.6 |     4.71 |
|             | Fixed         |                    31.45 |                       32.72 |     3.88 |
|             | Prefix        |                     72.6 |                        91.2 |    20.39 |
|             | Suffix        |                     64.8 |                        70.8 |     8.47 |
```

I am not sure why the performance for query 14 regressed with the changes, the data for 1000 runs computed earlier for Tpch table columns doesn't show a performance degradation. The Tpch benchmarks shared earlier were computed about two months ago and it needs to be checked again with the current PR.
```
| 1000 runs (time in seconds) |                    |                       |             | % Change |
|:---------------------------:|:------------------:|:---------------------:|:-----------:|----------|
|            Query            | With optimizations | Without optimizations | Savings (%) |     2.45 |
|              q2             |        15.16       |         24.84         |    38.97    |    -0.72 |
|              q9             |        18.39       |         22.37         |    17.79    |    20.64 |
|             q13             |        37.24       |          35.3         |     -5.5    |     4.95 |
|             q14             |        11.98       |         12.42         |     3.54    |    -2.99 |
|            q16_1            |        11.88       |         12.65         |     6.09    |     5.08 |
|            q16_2            |        32.57       |         33.49         |     2.75    |    21.24 |
|             q20             |        12.17       |         11.84         |    -2.79    |     6.96 |
|                        5000 | Wildcard only      |                 109.2 |       114.6 |     4.71 |
|                             | Fixed              |                 31.45 |       32.72 |     3.88 |
|                             | Prefix             |                  72.6 |        91.2 |    20.39 |
|                             | Suffix             |                  64.8 |        70.8 |     8.47 |
```