@mbasmanova 2022-09-06
@pramodsatya 

The following table is confusing to me. What are there two separate columns "Savings (%)" and "% Change". What values do they show?

I'm not seeing regression for q14. Time w/ opt is 11.98 and w/o 12.42. Where do you see the regression? q13 and q20 numbers do suggest a regression. Would you investigate? CC: @aditi-pandit 

> I am not sure why the performance for query 14 regressed with the changes, the data for 1000 runs computed earlier for Tpch table columns doesn't show a performance degradation. The Tpch benchmarks shared earlier were computed about two months ago and it needs to be checked again with the current PR.
```
| 1000 runs (time in seconds) |                    |                       |             | % Change |
|:---------------------------:|:------------------:|:---------------------:|:-----------:|----------|
|            Query            | With optimizations | Without optimizations | Savings (%) |     2.45 |
|              q2             |        15.16       |         24.84         |    38.97    |    -0.72 |
|              q9             |        18.39       |         22.37         |    17.79    |    20.64 |
|             q13             |        37.24       |          35.3         |     -5.5    |     4.95 |
|             q14             |        11.98       |         12.42         |     3.54    |    -2.99 |
|            q16_1            |        11.88       |         12.65         |     6.09    |     5.08 |
|            q16_2            |        32.57       |         33.49         |     2.75    |    21.24 |
|             q20             |        12.17       |         11.84         |    -2.79    |     6.96 |
```