@mbasmanova 2022-09-06
@pramodsatya q9 shows a gain of 17%, but is uses LIKE pattern that is not being optimized: `p.name like '%green%'`. How is this possible?