@mbasmanova 2022-09-06
@pramodsatya The regressions might be due to a new switch in the hot path. A fix could be to turn LikeConstantPattern into a template parameterized on PatternKind and move pattern analysis logic into makeLike.