@pramodsatya 2022-09-09
The benchmarks for Tpch queries having the like operator and patterns optimized for in this PR are as follows:
```
| No. of rows | Query Number | With optimization PR (s) | Without optimization PR (s) | % Change | 
|-------------|--------------|--------------------------|-----------------------------|----------|
|        1000 |            2 |                    13.78 |                       17.74 |    22.32 | 
|             |           14 |                    11.95 |                       12.49 |     4.32 |     
|             |           16 |                    12.12 |                       12.48 |     2.88 |     
|             |           20 |                    11.97 |                       12.57 |     4.77 |     
|        2000 |            2 |                    26.59 |                        32.2 |    17.42 |
|             |           14 |                     23.1 |                       23.72 |     2.61 |     
|             |           16 |                    23.46 |                       23.71 |     1.05 |    
|             |           20 |                    23.28 |                       23.36 |     0.34 |    
|        4000 |            2 |                    51.33 |                        63.6 |    19.29 | 
|             |           14 |                    46.09 |                       47.32 |      2.6 |      
|             |           16 |                    46.84 |                       47.15 |     0.66 |     
|             |           20 |                    46.09 |                       46.56 |     1.01 |     
|        8000 |            2 |                    104.4 |                       136.2 |    23.35 |
|             |           14 |                     91.8 |                        95.4 |     3.77 | 
|             |           16 |                       93 |                        95.4 |     2.52 | 
|             |           20 |                     92.4 |                        94.2 |     1.91 |
```
Query 2 shows the most improvement in performance at ~20%. Number of rows here is the vector size of the Tpch table generated with helper functions from TpchGen.cpp. It is better to have LikeConstantPattern templated on PatternKind as suggested earlier, instead of the switch case on patternKind in the match function:
```
| No. of rows | Query Number | With optimization PR (s) | Without optimization PR (s) | Without templated LikeConstantPattern (s) | 
|-------------|--------------|--------------------------|-----------------------------|-------------------------------------------|
|        8000 |            2 |                    104.4 |                       136.2 |                                     109.8 |    
|             |           14 |                     91.8 |                        95.4 |                                      97.8 |     
|             |           16 |                       93 |                        95.4 |                                     100.2 |    
|             |           20 |                     92.4 |                        94.2 |                                      91.8 |   
```